requirements.txt
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
diff-match-patch==20241021
Django==5.0.3
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.2
django-import-export==4.3.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
executing==2.0.1
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tablib==3.8.0
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2025.2
wcwidth==0.2.13
